pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parqueryd"
dynamic = ["version"]
description = "A distribution framework for parquery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GPL2"}
authors = [
    {name = "Carst Vaartjes", email = "cvaartjes@visualfabriq.com"}
]
maintainers = [
    {name = "Jelle Verstraaten", email = "jverstraaten@visualfabriq.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "azure-storage-blob>=12.4.0",
    "boto3>=1.17.95",
    "configobj>=5.0.6",
    "netifaces>=0.10.9",
    "pyarrow>=1.0.0",
    "parquery==2.0.2",
    "psutil>=5.7.2",
    "pyzmq==25.1.2",
    "redis>=3.5",
    "sentry-sdk",
    "smart-open>=1.11.1",
]

[project.optional-dependencies]
test = [
    "pytest>=4.6.11",
    "pytest-cov>=2.10.0",
    "codacy-coverage>=1.3.11",
    "moto",
    "numpy",
    "pandas>=1.5.3",
]
dev = [
    "ipython>=7.20",
    "ruff>=0.8.0",
]
all = [
    "pytest>=4.6.11",
    "pytest-cov>=2.10.0",
    "codacy-coverage>=1.3.11",
    "moto",
    "numpy",
    "pandas>=1.5.3",
    "ipython>=7.20",
    "ruff>=0.8.0",
]

[project.scripts]
parqueryd = "parqueryd.node:main"
parqueryd_local = "parqueryd.connect_local:main"

[project.urls]
Homepage = "https://github.com/visualfabriq/parqueryd"

[tool.setuptools]
packages = ["parqueryd"]
zip-safe = true

[tool.setuptools.dynamic]
version = {attr = "parqueryd.__version__"}

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function call in argument defaults
    "B904",   # within an except clause, raise exceptions with raise ... from err
]

[tool.ruff.lint.isort]
known-first-party = ["parqueryd"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
